requirements.txt
Flask==2.1.1
Werkzeug==2.0.3
opencv-python-headless==4.5.5.64
numpy==1.21.6
ultralytics
Pillow==9.1.1
